ScenTreesMakie.jl

Module Index

ScenTreesMakie.build_probabilities!

ScenTreesMakie.gaussian_path

ScenTreesMakie.height

ScenTreesMakie.kernel_scenarios

ScenTreesMakie.lat_plot

ScenTreesMakie.lattice_approximation

ScenTreesMakie.leaves

ScenTreesMakie.nodes

ScenTreesMakie.part_tree

ScenTreesMakie.plot_path!

ScenTreesMakie.root

ScenTreesMakie.running_maximum

ScenTreesMakie.sample_path

ScenTreesMakie.stage

ScenTreesMakie.structure

ScenTreesMakie.tree_approximation!

ScenTreesMakie.tree_plot

ScenTreesMakie.tree_plot!

Detailed API

ScenTreesMakie.build_probabilities! — Method

build_probabilities!(trr::Tree,probabilities::Array{Float64,2})

Returns the probabilities of the nodes without probabilities if the array of probabilities is less than the length of parents in the stochastic approximation procedure.

ScenTreesMakie.gaussian_path — Function

gaussian_path(n::Int64 = 4)

Returns a 'nx1' dimensional array of Gaussian random walk.

ScenTreesMakie.height — Method

height(trr::Tree)

Returns the height of the tree trr which is just the maximum number of the stages of each node.

ScenTreesMakie.kernel_scenarios — Function

kernel_scenarios(data::Union{Array{Int64,2},Array{Float64,2}}, kernelDistribution = Logistic; Markovian::Bool = true)

Returns an instance of a closure of the conditional density estimation method.

Args:

data - a matrix of data in 2 dimension (NxT) where N is the number of trajectories and T is the number of stages.

kernelDistribution - the distribution of the kernel functions. Default is Logistic distribution.

An optional keyword Markovian is passed to handle weights for Markovian trajectories or non-Markovian ones.

To get a Markov trajectory from above, KernelScenarios(data,kernelDistribution;Markovian=true)() gives the result for Markovian samples, which is used to generate a scenario lattice.

ScenTreesMakie.lat_plot — Function

lat_plot(lt::Lattice,title = nothing, label=nothing)

Returns a plot of a scenario lattice.

Arguments

lt - A scenario lattice.

title - A title for the plot.

simple - if simple is true, then unreachable paths are dropped, see details below.

label - optional labels with LaTeX support.

Value

- returns the plot as a gr() figure.

Details:

if simple is equal to true then states valued at 0.0 and have 0.0 probability to be reached are not plotted. 
To make use of this feature the `check_tree()` routine should be used before (SLOW!)

There might connections with zero probability –> these shouldn't be plotted! TODO

ScenTreesMakie.lattice_approximation — Function

lattice_approximation(bstructure::Array{Int64,1}, path::Function, nIterations::Int64, r::Int64 = 2, dim::Int64 = 1)

Returns a valuated approximated lattice for the stochastic process provided in any dimension. The default dimension is dim = 1 for a lattice in 1 dimension.

Args:

bstructure - Branching structure of the scenario lattice e.g., bstructure = [1,2,3,4,5] represents a 5-staged lattice

path - Function generating samples from a known distribution with length equal to the length of bstructure of the lattice.

nIterations - Number of iterations to be performed.

r - Parameter for the transportation distance.

dim - dimension of the lattice to be generated. This depends entirely on the dimension of the samples generated.

ScenTreesMakie.leaves — Method

leaves(trr::Tree,node=Int64[])

Returns the leaf nodes, their indexes and the conditional probabilities based on parent node.

Arguments:

trr - an instance of a Tree.

node - a node for which all leaf nodes are found.

Values

leaves - a vector of leaf nodes.

omegas - a vector of indizes indicating the order of the elements in leaves.

prob - the conditional probability of each leaf node conditional on node.

If node is not given, then node is treated as the root node.

ScenTreesMakie.nodes — Method

nodes(trr::Tree,t=Int64[])

Returns the nodes in the tree trr at stage t.

Example:

julia> trr = Tree(404);
julia> nodes(trr,1)
2-element Vector{Int64}:
 2
 3

ScenTreesMakie.part_tree — Function

part_tree(trr::Tree,node)

Returns a reduced tree starting at node. If node is the root node then original tree is returned.

Arguments:

trr - an instance of Tree

node - a node of trr which becomes the new root node.

ScenTreesMakie.plot_path! — Method

plot_path!(path)

Plots a given vector path on top of the tree/lattice

ScenTreesMakie.root — Method

root(trr::Tree,nodes=Int64[])

Returns the root of the tree if the node is not specified.

#Arguments:

trr - an instance of Tree.

nodes - target node from root.

If nodes is not specified, it returns the root of the tree.

If nodes is specified, it returns a sequence of nodes from the root to the specified node.

ScenTreesMakie.running_maximum — Function

running_maximum(n::Int64 = 4)

Returns a 'nx1' dimensional array of Running Maximum process.

ScenTreesMakie.sample_path — Method

sample_path(trr::Tree,nPath=1, flag_show=true)

Plots nPath sample paths of the tree trr weighted by the probability of the path occuring. TODO: plot whole tree, make opaque and only sample paths are dark

ScenTreesMakie.stage — Method

stage(trr::Tree, node=Int64[])

Returns the stage of each node in the tree.

Arguments:

trr - an instance of a Tree.

node - the number of node in the scenario tree you want to know its stage.

ScenTreesMakie.structure — Method

structure(lat::Lattice)

Provides a summary on the branching structure of the lattice lat.

Values

bs: vector indicating the number of nodes in each stage, i.e. the branching structure.

nodes: the total number of nodes.

edges: the total number of edges.

paths: the total number of unique paths.

Example

julia> lat = Lattice(304);
julia> structure(lat)
([1, 2, 3, 4], 10, 20, 24)

ScenTreesMakie.tree_approximation! — Method

tree_approximation!(newtree::Tree, path::Function, nIterations::Int64, p::Int64=2, r::Int64=2)

Returns a valuated probability scenario tree approximating the input stochastic process.

Args:

newtree - Tree with a certain branching structure,

path - function generating samples from the stochastic process to be approximated,

nIterations - number of iterations for stochastic approximation procedure,

p - choice of norm (default p = 2 (Euclidean distance)), and,

r - transportation distance parameter

ScenTreesMakie.tree_plot! — Method

tree_plot!(trr::Tree,title=nothing, simple = false, label=nothing)

Adds an additional tree plot to an existing plot using plot!(). See tree_plot() for details.

ScenTreesMakie.tree_plot — Method

tree_plot(trr::Tree,title=nothing, simple = false, label=nothing)

Returns the plot of the input tree trr.

Arguments

trr - A scenario tree.

title - A title for the plot.

simple - if simple is true, then unreachable paths are dropped, see details below.

label - optional labels with LaTeX support.

Value

- returns the plot as a gr() figure.

Details:

if simple is equal to true then states valued at 0.0 and have 0.0 probability to be reached are not plotted. 
To make use of this feature the `check_tree()` routine should be used before (SLOW!)